% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{assert_mutually_exclusive}
\alias{assert_mutually_exclusive}
\title{Assert that at most one of several arguments is supplied}
\usage{
assert_mutually_exclusive(..., call = rlang::caller_env())
}
\arguments{
\item{...}{Two or more arguments to check, typically optional parameters of
the calling function that default to \code{NULL}.}

\item{call}{Environment used as the error's call context. Defaults to the
calling function, so errors point at the user's code.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Treats a \code{NULL} argument as "not supplied". Raises an error if two or more
of the arguments in \code{...} are non-\code{NULL}. Supplying none is allowed.
}
\examples{
draw <- function(colour = NULL, palette = NULL) {
  assert_mutually_exclusive(colour, palette)
  "ok"
}
draw(colour = "red")

}
