% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{assert_null}
\alias{assert_null}
\title{Assert that an object is NULL}
\usage{
assert_null(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{Object to check.}

\item{arg}{Name used to refer to \code{x} in error messages. Defaults to the
name of the expression passed as \code{x}.}

\item{call}{Environment used as the error's call context. Defaults to the
calling function, so errors point at the user's code.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Assert that an object is NULL
}
\examples{
assert_null(NULL)

}
