% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Validate function arguments}
\usage{
assert(..., msg = "", on_fail = base::stop)
}
\arguments{
\item{...}{List of logical assertions}

\item{msg}{Nothing if all assertions pass. Otherwise throws an error describing which
assertions failed.}

\item{on_fail}{Function to use to stop excution. Should be set to base::stop or
base::warning}
}
\value{
Nothing if all assertions pass. Otherwise throws an error describing which
}
\description{
Assert that each of the provided expression is true. Otherwise stop execution and
return a description of each failed assertion.
}
\examples{
\dontrun{
assert(2 < 1, 1 == 1, "1" == 2)

sum <- function(a, b) {
  assert(is.numeric(a),
         is.numeric(b),
         length(a) == length(b))
  return(a+b)
}
sum("1", 2)
}

}
