% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{vector-assertions}
\alias{vector-assertions}
\alias{assert_character}
\alias{assert_numeric}
\alias{assert_integer}
\alias{assert_double}
\alias{assert_logical}
\alias{assert_complex}
\alias{assert_list}
\alias{assert_factor}
\alias{assert_function}
\title{Assert that an object is a vector of a given type}
\usage{
assert_character(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_numeric(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_integer(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_double(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_logical(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_complex(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_list(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_factor(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)

assert_function(
  x,
  null_ok = FALSE,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Object to check.}

\item{null_ok}{Single logical. If \code{TRUE}, a \code{NULL} value passes the check
without error. Use this for optional arguments that default to \code{NULL}.
Defaults to \code{FALSE}, so \code{NULL} is rejected unless you opt in.}

\item{arg}{Name used to refer to \code{x} in error messages. Defaults to the
name of the expression passed as \code{x}.}

\item{call}{Environment used as the error's call context. Defaults to the
calling function, so errors point at the user's code.}
}
\value{
The input \code{x}, invisibly.
}
\description{
Each function checks that \code{x} is of the named type, allowing any length.
To also constrain the length, follow with \code{\link[=assert_length]{assert_length()}}. On success
the input is returned invisibly; on failure an error is raised against the
calling function.
}
\examples{
assert_character(c("a", "b", "c"))
assert_numeric(1:10)
assert_list(list(1, "a", TRUE))

}
